move common user actions to common section
